\name{dtMacroalgae-package}
\alias{dtMacroalgae-package}
\alias{dtMacroalgae}
\docType{package}
\title{
  Model of macroalgae growth in marine environments, patterned to Phyllophora
}
\description{
  Contains dynamic descriptions of the growth of macroalgae in the environment.
  
  The model is implemented in fortran and linked to R.
  
    \if{html}{\figure{Macroalgae_logo.png}{options: width=80 alt="Logo" align='right'}}
\if{latex}{\figure{Macroalgae_logo}{options: width=0.4in}}
  
}

\author{
  Quiten Mudde, Karline Soetaert
}

\keyword{ package }
